Refactor SemanticModelBuilderVisitor into smaller visitors
Splits the monolithic SemanticModelBuilderVisitor into three focused visitors:
- ErrorVisitor: For syntax error checking.
- DeclarationVisitor: For pass 1 (creating skeleton objects).
- LinkingVisitor: For pass 2 (analyzing and linking logic).

This improves maintainability and separation of concerns as identified in the code quality report.
