+++ CODE_QUALITY_REPORT.md
@@ -42,7 +42,7 @@ The `daedalus-dialog-editor` and `daedalus-parser` project demonstrates a solid 
 ### Priority: Medium (Sustainability)
 1.  **Optimize Graph Rendering:** [Done] Memoize the result of `buildQuestGraph` or debounce the `semanticModel` updates passed to `QuestFlow`.
 2.  **Decouple Serialization:** Use a library like `zod` or `class-transformer` for serialization/deserialization to remove manual boilerplate in `semantic-model.ts`.
-3.  **Split Visitor Logic:** Refactor `SemanticModelBuilderVisitor` into smaller, focused visitors (e.g., `DeclarationVisitor`, `ReferenceLinkingVisitor`).
+3.  **Split Visitor Logic:** [Done] Refactor `SemanticModelBuilderVisitor` into smaller, focused visitors (e.g., `DeclarationVisitor`, `ReferenceLinkingVisitor`).
 
 ### Priority: Low (Long-term)
 1.  **Worker Threads:** Move parsing logic to a Worker Thread to unblock the main process.
